Reorganized unit tests by functional area
