+++ pallets/msa/src/tests/common_tests.rs
@@ -1,23 +1,17 @@
 use frame_support::{
 	assert_err, assert_noop, assert_ok,
-	dispatch::{DispatchInfo, GetDispatchInfo, Pays, Weight},
-	pallet_prelude::InvalidTransaction,
-	traits::{ChangeMembers, Hash},
+	dispatch::{GetDispatchInfo, Pays},
 	BoundedBTreeMap,
 };
 
-use sp_core::{crypto::AccountId32, sr25519, sr25519::Public, Encode, Pair};
-use sp_runtime::{
-	traits::SignedExtension, transaction_validity::TransactionValidity, ArithmeticError,
-	MultiSignature,
-};
+use sp_core::{crypto::AccountId32, sr25519, Encode, Pair};
+use sp_runtime::{ArithmeticError, MultiSignature};
 
 use crate::{
 	ensure,
-	mock::*,
+	tests::mock::*,
 	types::{AddKeyData, AddProvider, PermittedDelegationSchemas, EMPTY_FUNCTION},
-	CheckFreeExtrinsicUse, Config, CurrentMsaIdentifierMaximum, DispatchResult, Error, Event,
-	ProviderToRegistryEntry, ValidityError,
+	Config, DispatchResult, Error, Event, ProviderToRegistryEntry,
 };
 
 use common_primitives::{
@@ -29,29 +23,6 @@ use common_primitives::{
 	schema::{SchemaId, SchemaValidator},
 	utils::wrap_binary_data,
 };
-use common_runtime::extensions::check_nonce::CheckNonce;
-
-#[test]
-fn it_creates_an_msa_account() {
-	new_test_ext().execute_with(|| {
-		assert_ok!(Msa::create(test_origin_signed(1)));
-
-		assert_eq!(Msa::get_msa_by_public_key(test_public(1)), Some(1 as MessageSourceId));
-
-		assert_eq!(Msa::get_current_msa_identifier_maximum(), 1);
-
-		System::assert_last_event(Event::MsaCreated { msa_id: 1, key: test_public(1) }.into());
-	});
-}
-
-#[test]
-fn it_throws_msa_identifier_overflow() {
-	new_test_ext().execute_with(|| {
-		CurrentMsaIdentifierMaximum::<Test>::set(u64::MAX);
-
-		assert_noop!(Msa::create(test_origin_signed(1)), Error::<Test>::MsaIdOverflow);
-	});
-}
 
 #[test]
 #[allow(unused_must_use)]
@@ -62,16 +33,6 @@ fn it_does_not_allow_duplicate_keys() {
 		assert_noop!(Msa::create(test_origin_signed(1)), Error::<Test>::KeyAlreadyRegistered);
 
 		assert_eq!(Msa::get_current_msa_identifier_maximum(), 1);
-	});
-}
-
-#[test]
-fn it_create_has_weight() {
-	new_test_ext().execute_with(|| {
-		let call = MsaCall::<Test>::create {};
-		let dispatch_info = call.get_dispatch_info();
-
-		assert!(dispatch_info.weight.ref_time() > Weight::from_ref_time(10_000 as u64).ref_time());
 	});
 }
 
@@ -391,185 +352,6 @@ fn it_deletes_msa_key_successfully() {
 }
 
 #[test]
-fn test_retire_msa_success() {
-	new_test_ext().execute_with(|| {
-		let (test_account_key_pair, _) = sr25519::Pair::generate();
-
-		// Create an account
-		let test_account = AccountId32::new(test_account_key_pair.public().into());
-		let origin = RuntimeOrigin::signed(test_account.clone());
-
-		// Create an MSA so this account has one key associated with it
-		assert_ok!(Msa::create(origin.clone()));
-		let msa_id = Msa::get_owner_of(&test_account).unwrap();
-
-		// Retire the MSA
-		assert_ok!(Msa::retire_msa(origin));
-
-		// Check if PublicKeyDeleted event was dispatched.
-		System::assert_has_event(Event::PublicKeyDeleted { key: test_account.clone() }.into());
-
-		// Check if MsaRetired event was dispatched.
-		System::assert_last_event(Event::MsaRetired { msa_id }.into());
-
-		// Assert that the MSA has no accounts
-		let key_count = Msa::get_public_key_count_by_msa_id(msa_id);
-		assert_eq!(key_count, 0);
-
-		// MSA has been retired, perform additional tests
-
-		// [TEST] Adding an account to the retired MSA should fail
-		let (key_pair1, _) = sr25519::Pair::generate();
-		let new_account1 = key_pair1.public();
-		let (msa_id2, _) = create_account();
-
-		let add_new_key_data =
-			AddKeyData { msa_id: msa_id2, expiration: 10, new_public_key: new_account1.into() };
-
-		let encode_data_new_key_data = wrap_binary_data(add_new_key_data.encode());
-		let old_msa_owner_signature: MultiSignature =
-			test_account_key_pair.sign(&encode_data_new_key_data).into();
-		let new_key_signature: MultiSignature = key_pair1.sign(&encode_data_new_key_data).into();
-		assert_noop!(
-			Msa::add_public_key_to_msa(
-				RuntimeOrigin::signed(test_account.clone()),
-				test_account_key_pair.public().into(),
-				old_msa_owner_signature.clone(),
-				new_key_signature,
-				add_new_key_data
-			),
-			Error::<Test>::NoKeyExists
-		);
-
-		// [TEST] Adding a provider to the retired MSA should fail
-		let (provider_key_pair, _) = sr25519::Pair::generate();
-		let provider_account = provider_key_pair.public();
-
-		// Create provider account and get its MSA ID (u64)
-		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
-		let provider_msa_id =
-			Msa::ensure_valid_msa_key(&AccountId32::new(provider_account.0)).unwrap();
-
-		assert_ok!(Msa::create_provider(
-			RuntimeOrigin::signed(provider_account.into()),
-			Vec::from("Foo")
-		));
-
-		let (delegator_signature, add_provider_payload) =
-			create_and_sign_add_provider_payload(test_account_key_pair, provider_msa_id);
-
-		assert_noop!(
-			Msa::grant_delegation(
-				RuntimeOrigin::signed(provider_account.into()),
-				test_account.clone(),
-				delegator_signature,
-				add_provider_payload
-			),
-			Error::<Test>::NoKeyExists
-		);
-
-		// [TEST] Revoking a delegation (modifying permissions) should not do anything
-		assert_revoke_delegation_by_delegator_no_effect(test_account, provider_msa_id)
-	})
-}
-
-fn assert_revoke_delegation_by_delegator_no_effect(
-	test_account: AccountId32,
-	provider_msa_id: u64,
-) {
-	let event_count = System::event_count();
-	assert_ok!(Msa::revoke_delegation_by_delegator(
-		RuntimeOrigin::signed(test_account.clone()),
-		provider_msa_id
-	));
-	assert_eq!(event_count, System::event_count())
-}
-
-#[test]
-fn test_retire_msa_does_nothing_when_no_msa() {
-	new_test_ext().execute_with(|| {
-		let (test_pair, _) = sr25519::Pair::generate();
-		let first_account_key = test_pair.public();
-		let origin = RuntimeOrigin::signed(first_account_key.into());
-
-		// 1. when there's no MSA at all
-		let event_count = System::event_count();
-		assert_ok!(Msa::retire_msa(origin.clone()));
-		assert_eq!(event_count, System::event_count());
-	});
-}
-
-#[test]
-fn test_ensure_msa_can_retire_fails_if_registered_provider() {
-	new_test_ext().execute_with(|| {
-		// Create an account
-		let (test_account_key_pair, _) = sr25519::Pair::generate();
-		let test_account = AccountId32::new(test_account_key_pair.public().into());
-		let origin = RuntimeOrigin::signed(test_account.clone());
-
-		// Add an account to the MSA
-		assert_ok!(Msa::add_key(2, &test_account, EMPTY_FUNCTION));
-
-		// Register provider
-		assert_ok!(Msa::create_provider(origin, Vec::from("Foo")));
-
-		// Retire MSA
-		assert_noop!(
-			CheckFreeExtrinsicUse::<Test>::ensure_msa_can_retire(&test_account),
-			InvalidTransaction::Custom(
-				ValidityError::InvalidRegisteredProviderCannotBeRetired as u8
-			)
-		);
-	})
-}
-
-#[test]
-fn test_ensure_msa_can_retire_fails_if_more_than_one_account_exists() {
-	new_test_ext().execute_with(|| {
-		let msa_id = 2;
-		let (test_account_1_key_pair, _) = sr25519::Pair::generate();
-		let (test_account_2_key_pair, _) = sr25519::Pair::generate();
-
-		// Create accounts
-		let test_account_1 = AccountId32::new(test_account_1_key_pair.public().into());
-		let test_account_2 = AccountId32::new(test_account_2_key_pair.public().into());
-
-		// Add two accounts to the MSA
-		assert_ok!(Msa::add_key(msa_id, &test_account_1, EMPTY_FUNCTION));
-		assert_ok!(Msa::add_key(msa_id, &test_account_2, EMPTY_FUNCTION));
-
-		// Retire the MSA
-		assert_noop!(
-			CheckFreeExtrinsicUse::<Test>::ensure_msa_can_retire(&test_account_1),
-			InvalidTransaction::Custom(ValidityError::InvalidMoreThanOneKeyExists as u8)
-		);
-	})
-}
-
-#[test]
-fn test_ensure_msa_can_retire_fails_if_any_delegations_exist() {
-	new_test_ext().execute_with(|| {
-		// Create delegator
-		let msa_id = 2;
-		let (test_account_key_pair, _) = sr25519::Pair::generate();
-		let test_account = AccountId32::new(test_account_key_pair.public().into());
-		assert_ok!(Msa::add_key(msa_id, &test_account, EMPTY_FUNCTION));
-
-		// Create provider
-		let (provider_id, _provider_key) = create_provider_with_name("test");
-		let schema_ids = vec![1];
-		set_schema_count::<Test>(1);
-		assert_ok!(Msa::add_provider(ProviderId(provider_id), DelegatorId(msa_id), schema_ids));
-
-		// Retire the MSA
-		assert_err!(
-			CheckFreeExtrinsicUse::<Test>::ensure_msa_can_retire(&test_account),
-			InvalidTransaction::Custom(ValidityError::InvalidNonZeroProviderDelegations as u8)
-		);
-	})
-}
-
-#[test]
 pub fn test_get_owner_of() {
 	new_test_ext().execute_with(|| {
 		assert_eq!(Msa::get_owner_of(&test_public(1)), None);
@@ -664,287 +446,6 @@ pub fn add_provider_to_msa_is_success() {
 				provider_id: provider_msa.into(),
 			}
 			.into(),
-		);
-	});
-}
-
-#[test]
-pub fn grant_delegation_changes_schema_permissions() {
-	new_test_ext().execute_with(|| {
-		let (key_pair, _) = sr25519::Pair::generate();
-		let provider_account = key_pair.public();
-
-		let (delegator_pair, _) = sr25519::Pair::generate();
-		let delegator_account = delegator_pair.public();
-
-		// Create provider account and get its MSA ID (u64)
-		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
-		let provider_msa =
-			Msa::ensure_valid_msa_key(&AccountId32::new(provider_account.0)).unwrap();
-
-		// Create delegator account and get its MSA ID (u64)
-		assert_ok!(Msa::create(RuntimeOrigin::signed(delegator_account.into())));
-		let delegator_msa =
-			Msa::ensure_valid_msa_key(&AccountId32::new(delegator_account.0)).unwrap();
-
-		// Register provider
-		assert_ok!(Msa::create_provider(
-			RuntimeOrigin::signed(provider_account.into()),
-			Vec::from("Foo")
-		));
-
-		let block_expiration: BlockNumber = 110;
-
-		System::set_block_number(90);
-		set_schema_count::<Test>(10);
-
-		// Create delegation without any schema permissions
-		let (delegator_signature, add_provider_payload) =
-			create_and_sign_add_provider_payload_with_schemas(
-				delegator_pair.clone(),
-				provider_msa,
-				None,
-				block_expiration,
-			);
-
-		assert_ok!(Msa::grant_delegation(
-			RuntimeOrigin::signed(provider_account.into()),
-			delegator_account.into(),
-			delegator_signature,
-			add_provider_payload
-		));
-
-		let provider = ProviderId(provider_msa);
-		let delegator = DelegatorId(delegator_msa);
-
-		assert_eq!(
-			Msa::get_delegation(delegator, provider),
-			Some(Delegation { revoked_at: 0, schema_permissions: Default::default() })
-		);
-
-		System::assert_last_event(
-			Event::DelegationGranted {
-				delegator_id: delegator_msa.into(),
-				provider_id: provider_msa.into(),
-			}
-			.into(),
-		);
-
-		// Grant delegation w/schemas 1, 2, and 3 (implies block 0)
-		let (delegator_signature, add_provider_payload) =
-			create_and_sign_add_provider_payload_with_schemas(
-				delegator_pair.clone(),
-				provider_msa,
-				Some(vec![1, 2, 3]),
-				block_expiration,
-			);
-
-		assert_ok!(Msa::grant_delegation(
-			RuntimeOrigin::signed(provider_account.into()),
-			delegator_account.into(),
-			delegator_signature,
-			add_provider_payload
-		));
-
-		let mut sp = BoundedBTreeMap::<SchemaId, u64, MaxSchemaGrantsPerDelegation>::new();
-		assert_ok!(sp.try_insert(1u16, 0u64));
-		assert_ok!(sp.try_insert(2u16, 0u64));
-		assert_ok!(sp.try_insert(3u16, 0u64));
-
-		let expected = Delegation { revoked_at: 0, schema_permissions: sp };
-
-		assert_eq!(Msa::get_delegation(delegator, provider), Some(expected));
-
-		let revoked_block_number: u64 = 100;
-		System::set_block_number(revoked_block_number);
-		// Revoke all schema ids.
-		let (delegator_signature, add_provider_payload) =
-			create_and_sign_add_provider_payload_with_schemas(
-				delegator_pair.clone(),
-				provider_msa,
-				None,
-				block_expiration,
-			);
-
-		assert_ok!(Msa::grant_delegation(
-			RuntimeOrigin::signed(provider_account.into()),
-			delegator_account.into(),
-			delegator_signature,
-			add_provider_payload
-		));
-
-		let mut sp = BoundedBTreeMap::<SchemaId, u64, MaxSchemaGrantsPerDelegation>::new();
-		assert_ok!(sp.try_insert(1u16, revoked_block_number)); // schema id 1 revoked at revoked_block_number
-		assert_ok!(sp.try_insert(2u16, revoked_block_number)); // schema id 2 revoked at revoked_block_number
-		assert_ok!(sp.try_insert(3u16, revoked_block_number)); // schema id 3 revoked at revoked_block_number
-
-		let expected = Delegation { revoked_at: 0, schema_permissions: sp };
-
-		assert_eq!(Msa::get_delegation(delegator, provider), Some(expected));
-
-		System::set_block_number(revoked_block_number + 1);
-		// Grant 2, 3, 4
-		let (delegator_signature, add_provider_payload) =
-			create_and_sign_add_provider_payload_with_schemas(
-				delegator_pair.clone(),
-				provider_msa,
-				Some(vec![2, 3, 4]),
-				block_expiration,
-			);
-
-		assert_ok!(Msa::grant_delegation(
-			RuntimeOrigin::signed(provider_account.into()),
-			delegator_account.into(),
-			delegator_signature,
-			add_provider_payload
-		));
-
-		let mut sp = BoundedBTreeMap::<SchemaId, u64, MaxSchemaGrantsPerDelegation>::new();
-		assert_ok!(sp.try_insert(1u16, 100u64)); // schema id 1 revoked at block 100
-		assert_ok!(sp.try_insert(2u16, 0u64)); // schema id 2 granted (block 0)
-		assert_ok!(sp.try_insert(3u16, 0u64)); // schema id 3 granted (block 0)
-		assert_ok!(sp.try_insert(4u16, 0u64)); // schema id 4 granted (block 0)
-
-		let expected = Delegation { revoked_at: 0, schema_permissions: sp };
-		assert_eq!(Msa::get_delegation(delegator, provider), Some(expected));
-	});
-}
-
-#[test]
-pub fn grant_delegation_to_msa_throws_add_provider_verification_failed() {
-	new_test_ext().execute_with(|| {
-		let (key_pair, _) = sr25519::Pair::generate();
-		let account = key_pair.public();
-		let expiration: BlockNumber = 10;
-		let add_provider_payload = AddProvider::new(2, None, expiration);
-		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
-
-		let signature: MultiSignature = key_pair.sign(&encode_add_provider_data).into();
-		let fake_provider_payload = AddProvider::new(3, None, expiration);
-		assert_noop!(
-			Msa::grant_delegation(
-				RuntimeOrigin::signed(account.into()),
-				account.into(),
-				signature,
-				fake_provider_payload
-			),
-			Error::<Test>::AddProviderSignatureVerificationFailed
-		);
-	});
-}
-
-#[test]
-pub fn grant_delegation_throws_no_key_exist_error() {
-	new_test_ext().execute_with(|| {
-		let (key_pair, _) = sr25519::Pair::generate();
-		let provider_account = key_pair.public();
-
-		let expiration: BlockNumber = 10;
-		let add_provider_payload = AddProvider::new(2, None, expiration);
-		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
-
-		let signature: MultiSignature = key_pair.sign(&encode_add_provider_data).into();
-
-		assert_noop!(
-			Msa::grant_delegation(
-				test_origin_signed(1),
-				provider_account.into(),
-				signature,
-				add_provider_payload
-			),
-			Error::<Test>::NoKeyExists
-		);
-	});
-}
-
-#[test]
-pub fn grant_delegation_throws_key_revoked_error() {
-	new_test_ext().execute_with(|| {
-		let (key_pair, _) = sr25519::Pair::generate();
-		let provider_account = key_pair.public();
-
-		let expiration: BlockNumber = 10;
-		let add_provider_payload = AddProvider::new(2, None, expiration);
-		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
-
-		let signature: MultiSignature = key_pair.sign(&encode_add_provider_data).into();
-
-		assert_ok!(Msa::create(test_origin_signed(1)));
-		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
-		assert_ok!(Msa::delete_key_for_msa(1, &test_public(1)));
-
-		assert_noop!(
-			Msa::grant_delegation(
-				test_origin_signed(1),
-				provider_account.into(),
-				signature,
-				add_provider_payload
-			),
-			Error::<Test>::NoKeyExists
-		);
-	});
-}
-
-#[test]
-pub fn grant_delegation_throws_invalid_self_provider_error() {
-	new_test_ext().execute_with(|| {
-		let (key_pair, _) = sr25519::Pair::generate();
-		let provider_account = key_pair.public();
-
-		let expiration: BlockNumber = 10;
-		let add_provider_payload = AddProvider::new(1, None, expiration);
-		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
-		let signature: MultiSignature = key_pair.sign(&encode_add_provider_data).into();
-
-		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
-
-		assert_noop!(
-			Msa::grant_delegation(
-				RuntimeOrigin::signed(provider_account.into()),
-				provider_account.into(),
-				signature,
-				add_provider_payload
-			),
-			Error::<Test>::InvalidSelfProvider
-		);
-	});
-}
-
-#[test]
-pub fn grant_delegation_throws_unauthorized_delegator_error() {
-	new_test_ext().execute_with(|| {
-		// Generate a key pair for the provider
-		let (provider_key_pair, _) = sr25519::Pair::generate();
-		let provider_account = provider_key_pair.public();
-
-		// Generate a key pair for the delegator
-		let (delegator_key_pair, _) = sr25519::Pair::generate();
-		let delegator_account = delegator_key_pair.public();
-		assert_ok!(Msa::create(RuntimeOrigin::signed(delegator_account.into())));
-		let delegator_msa_id =
-			Msa::ensure_valid_msa_key(&AccountId32::new(delegator_account.0)).unwrap();
-
-		let expiration: BlockNumber = 10;
-		let add_provider_payload = AddProvider::new(delegator_msa_id, None, expiration);
-		let encode_add_provider_data = wrap_binary_data(add_provider_payload.encode());
-		let signature: MultiSignature = delegator_key_pair.sign(&encode_add_provider_data).into();
-
-		assert_ok!(Msa::create(RuntimeOrigin::signed(provider_account.into())));
-
-		// Register provider
-		assert_ok!(Msa::create_provider(
-			RuntimeOrigin::signed(provider_account.into()),
-			Vec::from("Foo")
-		));
-
-		assert_noop!(
-			Msa::grant_delegation(
-				RuntimeOrigin::signed(provider_account.into()),
-				delegator_account.into(),
-				signature,
-				add_provider_payload
-			),
-			Error::<Test>::UnauthorizedDelegator
 		);
 	});
 }
@@ -1288,6 +789,18 @@ pub fn revoke_provider_is_successful() {
 	});
 }
 
+pub fn assert_revoke_delegation_by_delegator_no_effect(
+	test_account: AccountId32,
+	provider_msa_id: u64,
+) {
+	let event_count = System::event_count();
+	assert_ok!(Msa::revoke_delegation_by_delegator(
+		RuntimeOrigin::signed(test_account.clone()),
+		provider_msa_id
+	));
+	assert_eq!(event_count, System::event_count())
+}
+
 #[test]
 fn revoke_delegation_by_delegator_does_nothing_when_no_msa() {
 	new_test_ext()
@@ -1562,155 +1075,6 @@ pub fn delegation_expired() {
 	})
 }
 
-/// Assert that revoking an MSA delegation passes the signed extension CheckFreeExtrinsicUse
-/// validation when a valid delegation exists.
-#[test]
-fn signed_extension_revoke_delegation_by_delegator_success() {
-	new_test_ext().execute_with(|| {
-		let (provider_msa_id, delegator_account) = create_provider_msa_and_delegator();
-		let call_revoke_delegation: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::revoke_delegation_by_delegator { provider_msa_id });
-		let info = DispatchInfo::default();
-		let len = 0_usize;
-		let result = CheckFreeExtrinsicUse::<Test>::new().validate(
-			&delegator_account.into(),
-			call_revoke_delegation,
-			&info,
-			len,
-		);
-		assert_ok!(result);
-	});
-}
-
-/// Assert that revoking an MSA delegation fails the signed extension CheckFreeExtrinsicUse
-/// validation when no valid delegation exists.
-#[test]
-fn signed_extension_fails_when_revoke_delegation_by_delegator_called_twice() {
-	new_test_ext().execute_with(|| {
-		let (provider_msa_id, delegator_account) = create_provider_msa_and_delegator();
-		let call_revoke_delegation: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::revoke_delegation_by_delegator { provider_msa_id });
-		let info = DispatchInfo::default();
-		let len = 0_usize;
-		let result = CheckFreeExtrinsicUse::<Test>::new().validate(
-			&delegator_account.into(),
-			call_revoke_delegation,
-			&info,
-			len,
-		);
-		assert_ok!(result);
-		assert_ok!(Msa::revoke_delegation_by_delegator(
-			RuntimeOrigin::signed(delegator_account.into()),
-			provider_msa_id
-		));
-
-		System::set_block_number(System::block_number() + 1);
-		let call_revoke_delegation: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::revoke_delegation_by_delegator { provider_msa_id });
-		let info = DispatchInfo::default();
-		let len = 0_usize;
-		let result_revoked = CheckFreeExtrinsicUse::<Test>::new().validate(
-			&delegator_account.into(),
-			call_revoke_delegation,
-			&info,
-			len,
-		);
-		assert!(result_revoked.is_err());
-	});
-}
-
-#[test]
-fn signed_extension_revoke_delegation_by_provider_success() {
-	new_test_ext().execute_with(|| {
-		let (delegator_msa_id, provider_account) = create_delegator_msa_and_provider();
-		let call_revoke_delegation: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::revoke_delegation_by_provider {
-				delegator: delegator_msa_id,
-			});
-		let info = DispatchInfo::default();
-		let len = 0_usize;
-		let result = CheckFreeExtrinsicUse::<Test>::new().validate(
-			&provider_account.into(),
-			call_revoke_delegation,
-			&info,
-			len,
-		);
-		assert_ok!(result);
-	})
-}
-
-fn assert_revoke_delegation_by_provider_err(
-	expected_err: InvalidTransaction,
-	provider_account: Public,
-	delegator_msa_id: u64,
-) {
-	let call_revoke_delegation: &<Test as frame_system::Config>::RuntimeCall =
-		&RuntimeCall::Msa(MsaCall::revoke_delegation_by_provider { delegator: delegator_msa_id });
-	let info = DispatchInfo::default();
-	let len = 0_usize;
-	let result = CheckFreeExtrinsicUse::<Test>::new().validate(
-		&provider_account.into(),
-		call_revoke_delegation,
-		&info,
-		len,
-	);
-	assert_err!(result, expected_err);
-}
-
-#[test]
-fn signed_extension_revoke_delegation_by_provider_fails_when_no_delegator_msa() {
-	new_test_ext().execute_with(|| {
-		let (_, provider_pair) = create_account();
-		let provider_account = provider_pair.public();
-
-		let delegator_msa_id = 33u64;
-		let expected_err = InvalidTransaction::Custom(ValidityError::InvalidDelegation as u8);
-		assert_revoke_delegation_by_provider_err(expected_err, provider_account, delegator_msa_id);
-	})
-}
-
-#[test]
-fn signed_extension_revoke_delegation_by_provider_fails_when_no_provider_msa() {
-	new_test_ext().execute_with(|| {
-		let (provider_pair, _) = sr25519::Pair::generate();
-		let provider_account = provider_pair.public();
-
-		let (delegator_msa, _) = create_account();
-
-		let expected_err = InvalidTransaction::Custom(ValidityError::InvalidMsaKey as u8);
-		assert_revoke_delegation_by_provider_err(expected_err, provider_account, delegator_msa);
-	});
-}
-
-#[test]
-fn signed_extension_revoke_delegation_by_provider_fails_when_no_delegation() {
-	new_test_ext().execute_with(|| {
-		let (_, provider_pair) = create_account();
-		let provider_account = provider_pair.public();
-		let (delegator_msa, _) = create_account();
-
-		let expected_err = InvalidTransaction::Custom(ValidityError::InvalidDelegation as u8);
-		assert_revoke_delegation_by_provider_err(expected_err, provider_account, delegator_msa);
-	})
-}
-
-/// Assert that a call that is not one of the matches passes the signed extension
-/// CheckFreeExtrinsicUse validation.
-#[test]
-fn signed_extension_validation_valid_for_other_extrinsics() {
-	let random_call_should_pass: &<Test as frame_system::Config>::RuntimeCall =
-		&RuntimeCall::Msa(MsaCall::create {});
-	let info = DispatchInfo::default();
-	let len = 0_usize;
-	let result = CheckFreeExtrinsicUse::<Test>::new().validate(
-		&test_public(1),
-		random_call_should_pass,
-		&info,
-		len,
-	);
-	assert_ok!(result);
-}
-
 #[test]
 pub fn delete_msa_public_key_call_has_correct_costs() {
 	new_test_ext().execute_with(|| {
@@ -1723,135 +1087,6 @@ pub fn delete_msa_public_key_call_has_correct_costs() {
 		let dispatch_info = call.get_dispatch_info();
 		assert_eq!(dispatch_info.pays_fee, Pays::No);
 	})
-}
-
-#[test]
-fn signed_extension_validation_delete_msa_public_key_success() {
-	new_test_ext().execute_with(|| {
-		let (msa_id, original_key_pair) = create_account();
-
-		let (new_key_pair, _) = sr25519::Pair::generate();
-		let new_key: AccountId32 = new_key_pair.public().into();
-		assert_ok!(Msa::add_key(msa_id, &new_key, EMPTY_FUNCTION));
-
-		let original_key: AccountId32 = original_key_pair.public().into();
-
-		// set up call for new key to delete original key
-		let call_delete_msa_public_key: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::delete_msa_public_key {
-				public_key_to_delete: original_key.clone(),
-			});
-
-		let info = DispatchInfo::default();
-		let len = 0_usize;
-		assert_ok!(CheckFreeExtrinsicUse::<Test>::new().validate(
-			&new_key,
-			call_delete_msa_public_key,
-			&info,
-			len,
-		));
-
-		// validate other direction
-		let call_delete_msa_public_key2: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::delete_msa_public_key { public_key_to_delete: new_key });
-		assert_ok!(CheckFreeExtrinsicUse::<Test>::new().validate(
-			&original_key,
-			call_delete_msa_public_key2,
-			&info,
-			len,
-		));
-	});
-}
-
-#[test]
-fn signed_extension_validate_fails_when_delete_msa_public_key_called_twice() {
-	new_test_ext().execute_with(|| {
-		let (owner_msa_id, owner_key_pair) = create_account();
-
-		let (new_key_pair, _) = sr25519::Pair::generate();
-		let new_key: AccountId32 = new_key_pair.public().into();
-		assert_ok!(Msa::add_key(owner_msa_id, &new_key, EMPTY_FUNCTION));
-
-		let call_delete_msa_public_key: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::delete_msa_public_key {
-				public_key_to_delete: owner_key_pair.public().into(),
-			});
-
-		// check that it's okay to delete the original key
-		assert_ok!(CheckFreeExtrinsicUse::<Test>::new().validate(
-			&new_key,
-			call_delete_msa_public_key,
-			&DispatchInfo::default(),
-			0_usize,
-		));
-
-		// new key deletes the old key
-		assert_ok!(Msa::delete_msa_public_key(
-			RuntimeOrigin::signed(new_key.clone()),
-			owner_key_pair.public().into()
-		));
-
-		assert_validate_key_delete_fails(
-			&new_key,
-			owner_key_pair.public().into(),
-			ValidityError::InvalidMsaKey,
-		);
-	});
-}
-
-#[test]
-fn signed_extension_validate_fails_when_delete_msa_public_key_called_on_only_key() {
-	new_test_ext().execute_with(|| {
-		let (_, original_pair) = create_account();
-		let original_key: AccountId32 = original_pair.public().into();
-
-		assert_validate_key_delete_fails(
-			&original_key,
-			original_key.clone(),
-			ValidityError::InvalidSelfRemoval,
-		)
-	})
-}
-
-#[test]
-fn signed_extension_validate_fails_when_delete_msa_public_key_called_by_non_owner() {
-	new_test_ext().execute_with(|| {
-		let (_, original_pair) = create_account();
-		let original_key: AccountId32 = original_pair.public().into();
-
-		let (_, non_owner_pair) = create_account();
-		let non_owner_key: AccountId32 = non_owner_pair.public().into();
-		assert_validate_key_delete_fails(
-			&non_owner_key,
-			original_key.clone(),
-			ValidityError::NotKeyOwner,
-		)
-	})
-}
-
-// Assert that CheckFreeExtrinsicUse::validate fails with `expected_err_enum`,
-// for the "delete_msa_public_key" call, given extrinsic caller = caller_key,
-// when attempting to delete `public_key_to_delete`
-fn assert_validate_key_delete_fails(
-	caller_key: &AccountId32,
-	public_key_to_delete: AccountId32,
-	expected_err_enum: ValidityError,
-) {
-	let call_delete_msa_public_key: &<Test as frame_system::Config>::RuntimeCall =
-		&RuntimeCall::Msa(MsaCall::delete_msa_public_key { public_key_to_delete });
-
-	let expected_err: TransactionValidity =
-		InvalidTransaction::Custom(expected_err_enum as u8).into();
-
-	assert_eq!(
-		CheckFreeExtrinsicUse::<Test>::new().validate(
-			&caller_key,
-			call_delete_msa_public_key,
-			&DispatchInfo::default(),
-			0_usize,
-		),
-		expected_err
-	);
 }
 
 /// Assert that when a key has been added to an MSA, that it my NOT be added to any other MSA.
@@ -1955,118 +1190,6 @@ fn add_removed_key_to_msa_pass() {
 }
 
 #[test]
-fn create_provider_via_governance_happy_path() {
-	new_test_ext().execute_with(|| {
-		let (_new_msa_id, key_pair) = create_account();
-
-		// Create the provider based on 1 yes vote by the council
-		assert_ok!(Msa::create_provider_via_governance(
-			RuntimeOrigin::from(pallet_collective::RawOrigin::Members(1, 1)),
-			key_pair.public().into(),
-			Vec::from("ACME Widgets")
-		));
-		// Confirm that the MSA is now a provider
-		assert!(Msa::is_registered_provider(_new_msa_id));
-	})
-}
-
-/// Test that a request to be a provider, makes the MSA a provider after the council approves it.
-#[test]
-fn propose_to_be_provider_happy_path() {
-	new_test_ext().execute_with(|| {
-		// Create a new MSA account and request that it become a provider
-		let (_new_msa_id, key_pair) = create_account();
-		_ = Msa::propose_to_be_provider(
-			RuntimeOrigin::signed(key_pair.public().into()),
-			Vec::from("ACME Widgets"),
-		);
-
-		// Find the Proposed event and get it's hash and index so it can be voted on
-		let proposed_events: Vec<(u32, Hash)> = System::events()
-			.iter()
-			.filter_map(|event| match event.event {
-				RuntimeEvent::Council(pallet_collective::Event::Proposed {
-					account: _,
-					proposal_index,
-					proposal_hash,
-					threshold: _,
-				}) => Some((proposal_index, proposal_hash)),
-				_ => None,
-			})
-			.collect();
-
-		assert_eq!(proposed_events.len(), 1);
-
-		let proposal_index = proposed_events[0].0;
-		let proposal_hash = proposed_events[0].1;
-		let proposal = Council::proposal_of(proposal_hash).unwrap();
-		let proposal_len: u32 = proposal.encoded_size() as u32;
-
-		// Set up the council members
-		let council_member = test_public(1); // Use ALICE as the council member
-
-		let incoming = vec![];
-		let outgoing = vec![];
-		Council::change_members(&incoming, &outgoing, vec![council_member.clone()]);
-
-		// Vote YES on the proposal
-		assert_ok!(Council::vote(
-			RuntimeOrigin::signed(council_member.clone()),
-			proposal_hash,
-			proposal_index,
-			true
-		));
-
-		// Find the Voted event and check if it passed
-		let voted_events: Vec<(bool, u32, u32)> = System::events()
-			.iter()
-			.filter_map(|event| match event.event {
-				RuntimeEvent::Council(pallet_collective::Event::Voted {
-					account: _,
-					proposal_hash: _,
-					voted,
-					yes,
-					no,
-				}) => Some((voted, yes, no)),
-				_ => None,
-			})
-			.collect();
-
-		assert_eq!(voted_events.len(), 1);
-		assert_eq!(voted_events[0].0, true); // Was it voted on?
-		assert_eq!(voted_events[0].1, 1); // There should be one YES vote to pass
-
-		// Close the voting
-		assert_ok!(Council::close(
-			RuntimeOrigin::signed(test_public(5)),
-			proposal_hash,
-			proposal_index,
-			Weight::MAX,
-			proposal_len
-		));
-
-		// Find the Closed event and check if it passed
-		let closed_events: Vec<(u32, u32)> = System::events()
-			.iter()
-			.filter_map(|event| match event.event {
-				RuntimeEvent::Council(pallet_collective::Event::Closed {
-					proposal_hash: _,
-					yes,
-					no,
-				}) => Some((yes, no)),
-				_ => None,
-			})
-			.collect();
-
-		assert_eq!(closed_events.len(), 1);
-		assert_eq!(closed_events[0].0, 1); // There should be one YES vote to pass
-
-		// Confirm that the MSA is now a provider
-		assert!(Msa::is_registered_provider(_new_msa_id));
-	})
-}
-
-#[test]
 fn create_provider() {
 	new_test_ext().execute_with(|| {
 		let (_new_msa_id, key_pair) = create_account();
@@ -2111,7 +1234,7 @@ fn create_provider_duplicate() {
 	})
 }
 
-fn set_schema_count<T: Config>(n: u16) {
+pub fn set_schema_count<T: Config>(n: u16) {
 	<T>::SchemaValidator::set_schema_count(n);
 }
 
@@ -2205,124 +1328,6 @@ pub fn schema_granted_success_rpc() {
 		let expected_schemas_granted = vec![1, 2];
 		let output_schemas: Vec<SchemaId> = schemas_granted.unwrap().unwrap();
 		assert_eq!(output_schemas, expected_schemas_granted);
-	})
-}
-
-// Assert that check nonce validation does not create a token account for delete_msa_public_key call.
-#[test]
-fn signed_ext_check_nonce_delete_msa_public_key() {
-	new_test_ext().execute_with(|| {
-		// Generate a key pair for MSA account
-		let (msa_key_pair, _) = sr25519::Pair::generate();
-		let msa_new_key = msa_key_pair.public();
-
-		let len = 0_usize;
-
-		// Test the delete_msa_public_key() call
-		let call_delete_msa_public_key: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::delete_msa_public_key {
-				public_key_to_delete: AccountId32::from(msa_new_key),
-			});
-		let info = call_delete_msa_public_key.get_dispatch_info();
-
-		// Call delete_msa_public_key() using the Alice account
-		let who = test_public(1);
-		assert_ok!(CheckNonce::<Test>(0).pre_dispatch(
-			&who,
-			call_delete_msa_public_key,
-			&info,
-			len
-		));
-
-		// Did the call create a token account?
-		let created_token_account: bool;
-		match frame_system::Account::<Test>::try_get(who) {
-			Ok(_) => {
-				created_token_account = true;
-			},
-			Err(_) => {
-				created_token_account = false;
-			},
-		};
-
-		// Assert that the call did not create a token account
-		assert_eq!(created_token_account, false);
-	})
-}
-
-// Assert that check nonce validation does not create a token account for revoke_delegation_by_delegator call.
-#[test]
-fn signed_ext_check_nonce_revoke_delegation_by_delegator() {
-	new_test_ext().execute_with(|| {
-		let (provider_msa_id, _) = create_provider_msa_and_delegator();
-
-		// We are testing the revoke_delegation_by_delegator() call.
-		let call_revoke_delegation_by_delegator: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::revoke_delegation_by_delegator { provider_msa_id });
-
-		let len = 0_usize;
-
-		// Get the dispatch info for the call.
-		let info = call_revoke_delegation_by_delegator.get_dispatch_info();
-
-		// Call revoke_delegation_by_delegator() using the Alice account
-		let who = test_public(1);
-		assert_ok!(CheckNonce::<Test>(0).pre_dispatch(
-			&who,
-			call_revoke_delegation_by_delegator,
-			&info,
-			len
-		));
-
-		// Did the call create a token account?
-		let created_token_account: bool;
-		match frame_system::Account::<Test>::try_get(who) {
-			Ok(_) => {
-				created_token_account = true;
-			},
-			Err(_) => {
-				created_token_account = false;
-			},
-		};
-
-		// Assert that the call did not create a token account
-		assert_eq!(created_token_account, false);
-	})
-}
-
-// Assert that check nonce validation does create a token account for a paying call.
-#[test]
-fn signed_ext_check_nonce_creates_token_account_if_paying() {
-	new_test_ext().execute_with(|| {
-		//  Test that a  "pays" extrinsic creates a token account
-		let who = test_public(1);
-		let len = 0_usize;
-		let pays_call_should_pass: &<Test as frame_system::Config>::RuntimeCall =
-			&RuntimeCall::Msa(MsaCall::create {});
-
-		// Get the dispatch info for the create() call.
-		let pays_call_should_pass_info = pays_call_should_pass.get_dispatch_info();
-
-		// Call create() using the Alice account
-		assert_ok!(CheckNonce::<Test>(0).pre_dispatch(
-			&who,
-			pays_call_should_pass,
-			&pays_call_should_pass_info,
-			len
-		));
-
-		// Did the call create a token account?
-		let created_token_account: bool;
-		match frame_system::Account::<Test>::try_get(who) {
-			Ok(_) => {
-				created_token_account = true;
-			},
-			Err(_) => {
-				created_token_account = false;
-			},
-		};
-		// Assert that the call created a token account
-		assert_eq!(created_token_account, true);
 	})
 }
