+++ pallets/msa/src/tests/mod.rs
@@ -1,0 +1,10 @@
+pub mod mock;
+
+mod common_tests;
+mod creation_tests;
+mod delegation_tests;
+mod governance_tests;
+mod replay_tests;
+mod retirement_tests;
+mod signature_registry_tests;
+mod signed_extension_tests;
